Presentation improvements
Now that I understand CSS and HTML better, I improved the layout of the TextFileExperiment.html document. This includes adding a 'center.main' (see CSS) and adding colours to a foreground and background. A customised font was chosen (Helvetica) and google fonts API was abandoned due to privacy concerns.  Finally, the source titles (h4) were made not bold and italic, to distinguish them better.