Added link functionality
Now the program is able to parse and automatically add links to the h1 h2 and h3 headers. This means anyone can take a link of a relevant section of the factsheet and show off that particular section. This will be necessary as the factsheet will grow to large lengths in the future, making navigability a key feature to add. Additionally, CSSFile was changed to hide links in titles, making their presence is less obtrusive to a reader. Finally, java code was made more efficient by not creating repetitive <body> tags.