Code made more efficient
As title says, the java is made more efficient by no longer making 'oneleveldeeper' method recursive, as '##o' was not necessary. However, '##c' is still necessary to denote when a bullet point list ought to end. Also, that method was altered to call another method that handles the meta and title tags specifically.