New functionality added for allowing images in factsheet
Style.css updated to allow images to be in the centre of a page.
Figures2-4 were changed to have names easily accessible to the java program.
Java code was adapted to recognise when to search for images and what images should be retrieved. It also made sure images were aligned to the centre of a page.
In the txt file, the images were also given alttext, for accessibility concerns.