New Process of Automatically updating files
If txt file updates, a batch file is run that automatically: runs the java program, and commits it. The latter part is still being developed. The former was done by making all BufferedReader/Writer constructors be forced to encode to UTF-8, which make issues with special characters not present in the command-line running of the program.